Update regression test results.
In some reason country names have changed: e.g. from "Republic of Finland"
to "Finland". About a month ago the change was opposite.

The problem is propably in smartmet-library-locus library where a name
search result is cached without keyword value of queryoptions. If the
cache contains multiple cache hits for a cache keyword the result may
vary. Now keyword of a queryoptions will be included into cache key.
Hopefully the regression test result are stable in the future.